feat: rename files for .ts
