+++ nodemon.json
@@ -1,0 +1,5 @@
+{
+  "watch": ["src", "./index.ts"],
+  "ext": ".ts, .js",
+  "ignore": []
+}
